Added keyboard functionality
When activated the dropdown can be used with the keyboard instead of with the mouse.

Up, down arrow change focused element, escape closes the dropdown, enter and space activate focused element.
When search is enabled and a single selection is active, when search returns a single match pressing enter in the search box will activate the matched option.