Documentation for select by group:
When items are grouped by property you can also specify an array of groups that you can use to select the items by.
The extra-settings property selectByGroups accepts an array of the values of the groups that you want to be selectable.
The naming will use the groupByTextProvider function to give them an actual label.